Technical Work
Here I included my Statistics 431 Final Paper, which was the replacement assignment for a final due to the initial outbreak of Covid-19 at the time. I conducted a statistical review of the research paper and looked at the metrics and techniques used. Later on, I did a project with two classmates, Philip Chau and Brian Wong, in my CIS 545 Big Data Analytics class. Attached are the analysis paper and the project code used to model Pennsylvania Department of Transportation Crash Data from 2000-2019.